LayoutTests/inspector-obsolete/styles/selector-list-expected.txt

Tests representation of selector lists in the protocol. Bug 103118.

Text
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
FOO bAr, #inspected, MOO>BAR, htML div, Foo~Moo, MOO  { (selector-list.html:19)
color: green;

[expanded] 
#inspected  { (selector-list.html:10)

[expanded] 
#InSpEcTeD  { (selector-list.html:7)

[expanded] 
#inspected  { (selector-list.html:4)

[expanded] 
FOO bAr, #inspected, MOO>BAR, htML div, Foo~Moo, MOO  { (selector-list.html:19)
/-- overloaded --/ color: green;

[expanded] 
div  { (user agent stylesheet)
display: block;
